fix: keep @/db out of the client bundle (cloudflare:workers resolve error)
The consolidation added exported plain helper functions that call getDb()
inside src/server-fns/ modules. TanStack Start's client compile strips
createServerFn handler bodies (and their imports) but cannot strip exported
plain functions, so the helpers kept `import { getDb } from "@/db"` alive in
the client graph — and @/db imports `cloudflare:workers`, which Vite cannot
resolve in the browser:

  [vite:import-analysis] Failed to resolve import "cloudflare:workers"
  from "src/db/index.ts"

Fix, following the src/server/ business-logic convention:
- getUserCompetitionRegistrationsForUser + getPendingTeamInvitesForEmail
  move to src/server/competition-detail.ts (import "server-only")
- getPublicCompetitionDivisionsForCompetition moves to
  src/server/competition-divisions.ts (import "server-only")
- parseCompetitionSettings, PENDING_PURCHASE_MAX_AGE_MINUTES, and the public
  division types move to client-safe src/utils/competition-settings.ts so the
  server modules can use them without importing a server-fn file (no cycle);
  competition-divisions-fns.ts re-exports them so existing importers
  (including client routes) are unchanged
- server-fn files now reference the helpers only inside handlers, which the
  compiler strips from client output

Verified with a full `pnpm build` (client bundle resolves), type-check,
biome, and the pending-invites/heats test files.
